+++ svm-predictor/src/main/java/svm/predictor/data/retrieving/MoneyLineDataRetriever.java
@@ -21,8 +21,8 @@ public class MoneyLineDataRetriever extends BaseDataRetriever {
 	}
 
 	@Override
-	public Integer getLabel(GameInfoDto game) {
-		Integer label = (game.getHomeTeamPoints() >= game.getAwayTeamPoints()) ? 1 : -1;
+	public Double getLabel(GameInfoDto game) {
+		Double label = (game.getHomeTeamPoints() >= game.getAwayTeamPoints()) ? 1.0 : -1.0;
 		return label;
 	}
